observatorio_editorial/requirements_django.txt
# Dependencias para el Observatorio Editorial con Django
Django==5.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
django-cors-headers==4.3.1
typing-extensions==4.8.0
django-environ==0.11.2
djangorestframework==3.14.0